requirements.txt
Django==4.2

django-configurations==2.4.1
django-model-utils==4.3.1
django-picklefield==3.1
django-storages==1.13.2

boto3==1.26.110
docutils==0.19
embedly2==0.6.0

Markdown==3.4.3
pytils==0.4.1
Pillow==9.5.0
psycopg2==2.9.6
redis==4.5.4
requests==2.28.2

sentry-sdk==1.19.1

whitenoise==6.4.0

# development
pytest==7.3.0
pytest-django==4.5.2
responses==0.23.1

black==23.3.0
flake8==6.0.0
isort==5.12.0
